nebula_gui_flask/requirements.txt
Flask
Flask-SocketIO
eventlet
requests
websocket-client
python-multipart